feat: initial POC
Adds everything needed to get started on this. Includes a new go
application (entrypoint in cmd/api/main.go) that will start up an
echo http server. The http server is simply gonna log every request
it receives. It includes a couple of pieces of custom middleware to
check the content type, and enable some level of authentication.

The authentication just reads a key from the environment and verifies
that each request includes this key in an X-Auth-Key header; very
basic but should be enough for the current purposes.

Adds a makefile for managing the development environment, has all
the tasks needed to get started, but primarily just run `make dcup`
and everything should be up and running and you can hit the api on
localhost:3030.

The dev env uses air to handle hot recompilation when the code
changes, there is an air.toml config included and a script that
is used to run this in the app service.

Sets up a trivy configuration to check for vulnerabilities.