build: setup build and test for PR's
Unignores go.sum, don't know why i added this to .gitignore tbh.

Downgrades go to 1.25.4, due to issues with the toolchain for test
coverage in CI not working on 1.25.5.

Adds a goreleaser file to build the artifacts needed for each PR,
really just validating that the build is working correctly.

Adds a .trivyignore file, even though it's empty, will be referenced
from the trivy CI check so easier to just put it in now.

Adds a testcoverage config, that is used in the CI check to generate
a testcoverage figure.

Adds ci workflows for running a formatting check across the code.

Adds a workflow to run tests, even though there aren't any right
now.